Add Lazy and extract Reader navigation to ReaderPresenter (new)
