+++ WordPress/Classes/System/SplitViewRootPresenter.swift
@@ -3,6 +3,7 @@ import Combine
 import SwiftUI
 import WordPressAuthenticator
 import WordPressUI
+import WordPressShared
 
 /// The presenter that uses triple-column navigation for `.regular` size classes
 /// and a tab-bar based navigation for `.compact` size class.
@@ -15,15 +16,14 @@ final class SplitViewRootPresenter: RootViewPresenter {
 
     private var siteContent: SiteSplitViewContent?
     private var notificationsContent: NotificationsSplitViewContent?
-    private var readerContent: ReaderPresenter?
+    @Lazy private var readerPresenter = ReaderPresenter()
     private var welcomeContent: WelcomeSplitViewContent?
 
     private var displayingContent: SplitViewDisplayable? {
-        let possibleContent: [SplitViewDisplayable?] = [siteContent, notificationsContent, readerContent, welcomeContent]
+        let possibleContent: [SplitViewDisplayable?] = [siteContent, notificationsContent, $readerPresenter.value, welcomeContent]
         let displaying = possibleContent
             .compactMap { $0 }
             .filter { $0.isDisplaying(in: splitVC) }
-
         wpAssert(displaying.count <= 1)
 
         return displaying.first
@@ -119,12 +119,7 @@ final class SplitViewRootPresenter: RootViewPresenter {
                 content = notificationsContent!
             }
         case .reader:
-            if let readerContent {
-                content = readerContent
-            } else {
-                readerContent = ReaderPresenter()
-                content = readerContent!
-            }
+            content = readerPresenter
         }
 
         display(content: content)
@@ -292,8 +287,8 @@ final class SplitViewRootPresenter: RootViewPresenter {
         } else {
             sidebarViewModel.selection = .reader
             if let path {
-                wpAssert(readerContent != nil)
-                readerContent?.navigate(to: path)
+                wpAssert($readerPresenter.value != nil)
+                readerPresenter.navigate(to: path)
             }
         }
     }
@@ -357,7 +352,7 @@ extension SplitViewRootPresenter: UISplitViewControllerDelegate {
                 tabBarVC.showBlogDetails(for: blog)
             }
         case .reader:
-            if let selection = readerContent?.sidebar.viewModel.selection {
+            if let selection = readerPresenter.sidebar.viewModel.selection {
                 switch selection {
                 case .main(let readerStaticScreen):
                     switch readerStaticScreen {
